+++ components/review/CommentForm.tsx
@@ -15,7 +15,7 @@ import { Input } from "@/components/ui/input";
 import { commentType } from "@/constants";
 import React, { useRef } from "react";
 import { setComment } from "@/actions/comment.action";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import { usePathname } from "next/navigation";
 
 // フォームのバリデーションスキーマを定義
