+++ components/top/CreateReview.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import React from 'react'
-import { Button } from './ui/button'
+import { Button } from '../ui/button'
 import { useRouter } from 'next/navigation'
 
 const CreateReview = () => {
